refactor(prisma): move output folder under `@lib`
