+++ prisma-next.config.ts
@@ -2,7 +2,8 @@ import 'dotenv/config';
 import { defineConfig } from '@prisma-next/mongo/config';
 
 export default defineConfig({
-  contract: "./prisma/contract.ts",
+  contract: "./src/lib/prisma/contract.ts",
+  outputPath: "./src/lib/prisma",
   db: {
     connection: process.env['MONGODB_URI']!,
   },
