+++ src/lib/models/reaction.ts
@@ -1,6 +1,6 @@
 import type { InferRootRow } from '@prisma-next/mongo-orm';
-import type { Contract } from '../../../prisma/contract.d';
-import { db } from '../../../prisma/db';
+import type { Contract } from '@lib/prisma/contract.d';
+import { db } from '@lib/prisma/db';
 import type { ReactionsEnum } from './enums';
 
 export type IReaction = Omit<InferRootRow<Contract, 'Reaction'>, 'reaction'> & {
